:art: Separate type effect of websocket from browser/dom/mod.ts
