+++ browser/dom/caret.ts
