+++ browser/dom/ensure.ts
