+++ browser/dom/isHeightViewable.ts
