+++ browser/dom/mod.ts
@@ -1,0 +1,9 @@
+export * from "./node.ts";
+export * from "./motion.ts";
+export * from "./edit.ts";
+export * from "./press.ts";
+export * from "./click.ts";
+export * from "./statusBar.ts";
+export * from "./caret.ts";
+export * from "./dom.ts";
+export * from "./openInTheSameTab.ts";
