+++ browser/dom/statusBar.ts
