+++ browser/mod.ts
@@ -1,12 +1,2 @@
-export * from "./node.ts";
-export * from "./motion.ts";
-export * from "./edit.ts";
-export * from "./press.ts";
-export * from "./click.ts";
-export * from "./statusBar.ts";
-export * from "./caret.ts";
-export * from "./dom.ts";
-export * from "./openInTheSameTab.ts";
-export * from "./websocket/room.ts";
-export * from "./websocket/shortcuts.ts";
-export * from "./websocket/stream.ts";
+export * from "./dom/mod.ts";
+export * from "./websocket/mod.ts";
